IST Dataset

The International Stroke Trial Dataset

Usage

IST

Format

A data frame with 19435 rows and 112 columns:

AGE

Randomisation data: Age in years

CMPLASP

Other data and derived variables: Compliant for aspirin

CMPLHEP

Other data and derived variables: Compliant for heparin

CNTRYNUM

Other data and derived variables: Country code

COUNTRY

Other data and derived variables: Abbreviated country code

DALIVE

Recurrent stroke within 14 days: Discharged alive from hospital

DALIVED

Recurrent stroke within 14 days: Date Discharged alive from hospital

DAP

Data collected on 14 day/discharge form about treatments given in hospital: Non trial antiplatelet drug (Y/N)

DASP14

Data collected on 14 day/discharge form about treatments given in hospital: Aspirin given for 14 days or till death or discharge (Y/N)

DASPLT

Data collected on 14 day/discharge form about treatments given in hospital: Discharged on long term aspirin (Y/N)

DAYLOCAL

Randomisation data: Estimate of local day of week (assuming RDATE is Oxford)

DCAA

Data collected on 14 day/discharge form about treatments given in hospital: Calcium antagonists (Y/N)

DCAREND

Data collected on 14 day/discharge form about treatments given in hospital: Carotid surgery (Y/N)

DDEAD

Other events within 14 days: Other events within 14 days: Dead on discharge form

DDEADC

Other events within 14 days: Cause of death (1-Initial stroke/2-Recurrent stroke (ischaemic or unknown /3-Recurrent stroke (haemorrhagic)/4-Pneumonia /5-Coronary heart disease/6-Pulmonary embolism /7-Other vascular or unknown/8-Non-vascular/0-unknown)

DDEADD

Date of dead on discharge form (yyyy/mm/dd); NOTE: this death is not necessarily within 14 days of randomisation

DDEADX

Other events within 14 days: Comment on death

DDIAGHA

Final diagnosis of initial event: Haemorrhagic stroke

DDIAGISC

Final diagnosis of initial event: Ischaemic stroke

DDIAGUN

Final diagnosis of initial event: Indeterminate stroke

DEAD1

Indicator variables for specific causes of death: Initial stroke

DEAD2

Indicator variables for specific causes of death: Reccurent ischaemic/unknown stroke

DEAD3

Indicator variables for specific causes of death: Reccurent haemorrhagic stroke

DEAD4

Indicator variables for specific causes of death: Pneumonia

DEAD5

Indicator variables for specific causes of death: Coronary heart disease

DEAD6

Indicator variables for specific causes of death: Pulmonary embolism

DEAD7

Indicator variables for specific causes of death: Other vascular or unknown

DEAD8

Indicator variables for specific causes of death: Non vascular

DGORM

Data collected on 14 day/discharge form about treatments given in hospital: Glycerol or manitol (Y/N)

DHAEMD

Data collected on 14 day/discharge form about treatments given in hospital: Haemodilution (Y/N)

DHH14

Data collected on 14 day/discharge form about treatments given in hospital: Medium dose heparin given for 14 days etc in pilot (combine with above)

DIED

Other data and derived variables: Indicator variable for death (1=died; 0=did not die)

DIVH

Data collected on 14 day/discharge form about treatments given in hospital: Non trial intravenous heparin (Y/N)

DLH14

Data collected on 14 day/discharge form about treatments given in hospital: Low dose heparin given for 14 days or till death/discharge (Y/N)

DMAJNCH

Data collected on 14 day/discharge form about treatments given in hospital: Major non-cerebral haemorrhage (Y/N)

DMAJNCHD

Data collected on 14 day/discharge form about treatments given in hospital:

DMAJNCHX

Data collected on 14 day/discharge form about treatments given in hospital:

DMH14

Data collected on 14 day/discharge form about treatments given in hospital: Date of Major non-cerebral haemorrhage (yyyy/mm/dd)

DNOSTRK

Final diagnosis of initial event: Not a stroke

DNOSTRKX

Final diagnosis of initial event: Comment on Not a stroke

DOAC

Data collected on 14 day/discharge form about treatments given in hospital: Other anticoagulants (Y/N)

DPE

Other events within 14 days: Pulmonary embolism

DPED

Other events within 14 days: Date of Pulmonary embolism (yyyy/mm/dd)

DPLACE

Other events within 14 days: Discharge destination (A-Home /B-Relatives home /C-Residential care /D-Nursing home /E-Other hospital departments /U-Unknown)

DRSH

Recurrent stroke within 14 days: Haemorrhagic stroke

DRSHD

Recurrent stroke within 14 days: Date of Haemorrhagic stroke (yyyy/mm/dd)

DRSISC

Recurrent stroke within 14 days: Ischaemic recurrent stroke

DRSISCD

Recurrent stroke within 14 days: Date of Ischaemic recurrent stroke (yyyy/mm/dd)

DRSUNK

Recurrent stroke within 14 days: Unknown type

DRSUNKD

Recurrent stroke within 14 days: Date of Unknown type (yyyy/mm/dd)

DSCH

Data collected on 14 day/discharge form about treatments given in hospital: Non trial subcutaneous heparin (Y/N)

DSIDE

Data collected on 14 day/discharge form about treatments given in hospital: Other side effect (Y/N)

DSIDED

Data collected on 14 day/discharge form about treatments given in hospital: Date of Other side effect

DSIDEX

Data collected on 14 day/discharge form about treatments given in hospital: Comment of Other side effect

DSTER

Data collected on 14 day/discharge form about treatments given in hospital: Steroids (Y/N)

DTHROMB

Data collected on 14 day/discharge form about treatments given in hospital: Thrombolysis (Y/N)

DVT14

Indicator variables for specific causes of death: Indicator of deep vein thrombosis on discharge form

EXPD14

Other data and derived variables: Predicted probability of death at 14 days

EXPD6

Other data and derived variables: Predicted probability of death at 6 month

EXPDD

Other data and derived variables: Predicted probability of death/dependence at 6 month

FAP

Data collected at 6 months: On antiplatelet drugs

FDEAD

Data collected at 6 months: Dead at six month follow-up (Y/N)

FDEADC

Data collected at 6 months: Cause of death (1-Initial stroke /2-Recurrent stroke (ischaemic or unknown) /3-Recurrent stroke (haemorrhagic) /4-Pneumonia /5-Coronary heart disease /6-Pulmonary embolism /7-Other vascular or unknown /8-Non-vascular /0-unknown)

FDEADD

Data collected at 6 months: Date of death; NOTE: this death is not necessarily within 6 months of randomisation

FDEADX

Data collected at 6 months: Comment on death

FDENNIS

Data collected at 6 months: Dependent at 6 month follow-up (Y/N)

FLASTD

Data collected at 6 months: Date of last contact

FOAC

Data collected at 6 months: On anticoagulants

FPLACE

Data collected at 6 months: Place of residance at 6 month follow-up ( A-Home /B-Relatives home /C-Residential care /D-Nursing home /E-Other hospital departments /U-Unknown)

FRECOVER

Data collected at 6 months: Fully recovered at 6 month follow-up (Y/N)

FU1_COMP

Other data and derived variables: Date discharge form completed

FU1_RECD

Other data and derived variables: Date discharge form received

FU2_DONE

Other data and derived variables: Date 6 month follow-up done

H14

ndicator variables for specific causes of death: Cerebral bleed/heamorrhagic stroke within 14 days; this is slightly wider definition than DRSH an is used for analysis of cerebral bleeds

HOSPNUM

Randomisation data: Hospital number

HOURLOCAL

Randomisation data: Local time – hours

HTI14

Indicator variables for specific causes of death: Indicator of haemorrhagic transformation within 14 days

ID14

Other data and derived variables: Indicator of death at 14 days

ISC14

Indicator variables for specific causes of death: Indicator of ischaemic stroke within 14 days

MINLOCAL

Randomisation data: Local time – minutes

NCB14

Indicator variables for specific causes of death: Indicator of any non-cerebral bleed within 14 days

NCCODE

Other data and derived variables: Coding of compliance (see Table 3) doi:10.1186/1745-6215-13-24

NK14

Indicator variables for specific causes of death: Indicator of indeterminate stroke within 14 days

OCCODE

Other data and derived variables: Six month outcome ( 1-dead /2-dependent /3-not recovered /4-recovered /8 or 9 – missing status

ONDRUG

Data collected on 14 day/discharge form about treatments given in hospital: Estimate of time in days on trial treatment

PE14

Indicator variables for specific causes of death: Indicator of pulmonary embolism within 14 days

RASP3

Randomisation data: Aspirin within 3 days prior to randomisation (Y/N)

RATRIAL

Randomisation data: Atrial fibrillation (Y/N); not coded for pilot phase - 984 patients

RCONSC

Randomisation data: Conscious state at randomisation (F - fully alert, D - drowsy, U - unconscious)

RCT

Randomisation data: CT before randomisation (Y/N)

RDATE

Randomisation data: Date of randomisation

RDEF1

Randomisation data: Face deficit (Y/N/C=can't assess)

RDEF2

Randomisation data: Arm/hand deficit (Y/N/C=can't assess)

RDEF3

Randomisation data: Leg/foot deficit (Y/N/C=can't assess)

RDEF4

Randomisation data: Dysphasia (Y/N/C=can't assess)

RDEF5

Randomisation data: Hemianopia (Y/N/C=can't assess)

RDEF6

Randomisation data: Visuospatial disorder (Y/N/C=can't assess)

RDEF7

Randomisation data: Brainstem/cerebellar signs (Y/N/C=can't assess)

RDEF8

Randomisation data: Other deficit (Y/N/C=can't assess)

RDELAY

Randomisation data: Delay between stroke and randomisation in hours

RHEP24

Randomisation data: Heparin within 24 hours prior to randomisation (Y/N)

RSBP

Randomisation data: Systolic blood pressure at randomisation (mmHg)

RSLEEP

Randomisation data: Symptoms noted on waking (Y/N)

RVISINF

Randomisation data: Infarct visible on CT (Y/N)

RXASP

Randomisation data: Trial aspirin allocated (Y/N)

RXHEP

Randomisation data: Trial heparin allocated (M/L/N) \[M is coded as H=high in pilot\]

SET14D

Other data and derived variables: Know to be dead or alive at 14 days (1=Yes, 0=No); this does not necessarily mean that we know outcome at 6 monts – see OCCODE for this

SEX

Randomisation data: M=male; F=female

STRK14

Indicator variables for specific causes of death: Indicator of any stroke within 14 days

STYPE

Randomisation data: Stroke subtype (TACS/PACS/POCS/LACS/other

TD

Other data and derived variables: Time of death or censoring in days

TRAN14

Indicator variables for specific causes of death: Indicator of major non-cerebral bleed within 14 days

...

Details

Obtained from Sandercock, Peter; Niewada, Maciej; Czlonkowska, Anna. (2011). International Stroke Trial database (version 2), [dataset]. University of Edinburgh. Department of Clinical Neurosciences. doi:10.7488/ds/104 Under ODC-by licence

References

doi:10.7488/ds/104

Author

Sandercock P et al. Peter.Sandercock@ed.ac.uk