Import a correlation matrix

Imports a correlation matrix from a csv file generated by export_empty_cor_matrix

Usage

import_cor_matrix(file_path = "./correlation_matrix.csv")

Arguments

file_path

A path to the csv file, Default: './correlation_matrix.csv'

Value

a matrix of correlations that can be used with synthesise_data

Details

A function to import the user specified correlations generated from the csv file exported by the export_empty_cor_matrix function. Correlations should be entered into the CSV file, using rank order correlations. The correlation matrix should be symmetric and be positive semi definite.

See also

Examples

if (FALSE) { # \dontrun{
  import_cor_matrix("correlation_matrix.csv")
} # }