Import Marginal Distributions

Import the marginal distribution as exported from a Trusted Research Environment (TRE)

Usage

import_marginal_distributions(
  folder_path = ".",
  binary_variables_file = "",
  categorical_variables_file = "",
  continuous_variables_file = "",
  continuous_quantiles_file = "",
  summary_file = "summary.csv"
)

Arguments

folder_path

Where the marginal distribution files are located, Default: '.' see details.

binary_variables_file

filename for the binary_variables file, Default: ” see details.

categorical_variables_file

filename for the categorical variables file , Default: ” see details.

continuous_variables_file

filename for the continuous variables file, Default: ” see details.

continuous_quantiles_file

filename for the continuous quantiles file, Default: ” see details.

summary_file

filename for the summary file, Default: 'summary.csv' see details.

Value

Returns an object of a RESIDE class

Details

This function will import marginal distributions as generated within a Trusted Research Environment (TRE) using the function export_marginal_distributions(), the folder_path allows the path of the files provided by the TRE to be imported, this will default to the current working directory. The file parameters will provide the default file names if no filenames are specified.

See also

Examples

if (FALSE) { # \dontrun{
if(interactive()){
   marginals <- import_marginal_distributions()
 }
} # }